Move all filter logic to stories module
