+++ code/core/src/manager-api/modules/layout.ts
@@ -1,28 +1,21 @@
-import { SET_CONFIG, STORY_INDEX_INVALIDATED } from 'storybook/internal/core-events';
+import { SET_CONFIG } from 'storybook/internal/core-events';
 import type {
   API_Layout,
   API_LayoutCustomisations,
   API_PanelPositions,
-  API_PreparedIndexEntry,
   API_UI,
-  FilterFunction,
-  Tag,
-  TagsOptions,
 } from 'storybook/internal/types';
 
 import { global } from '@storybook/global';
 
 import { pick, toMerged } from 'es-toolkit/object';
 import { isEqual as deepEqual } from 'es-toolkit/predicate';
-import memoize from 'memoizerific';
 import type { ThemeVars } from 'storybook/theming';
 import { create } from 'storybook/theming/create';
 
-import { Tag as TagEnum } from '../../shared/constants/tags';
 import merge from '../lib/merge';
 import type { ModuleFn } from '../lib/types';
 import type { State } from '../root';
-import type Store from '../store';
 
 const { document } = global;
 
@@ -41,24 +34,6 @@ export interface SubState {
   selectedPanel: string | undefined;
   theme: ThemeVars;
 }
-
-const TAGS_FILTER = 'tags-filter';
-
-const BUILT_IN_FILTERS = {
-  _docs: (entry: API_PreparedIndexEntry, excluded?: boolean) =>
-    excluded ? entry.type !== 'docs' : entry.type === 'docs',
-  _play: (entry: API_PreparedIndexEntry, excluded?: boolean) =>
-    excluded
-      ? entry.type !== 'story' || !entry.tags?.includes(TagEnum.PLAY_FN)
-      : entry.type === 'story' && !!entry.tags?.includes(TagEnum.PLAY_FN),
-  _test: (entry: API_PreparedIndexEntry, excluded?: boolean) =>
-    excluded
-      ? entry.type !== 'story' || entry.subtype !== 'test'
-      : entry.type === 'story' && entry.subtype === 'test',
-};
-
-const USER_TAG_FILTER = (tag: Tag) => (entry: API_PreparedIndexEntry, excluded?: boolean) =>
-  excluded ? !entry.tags?.includes(tag) : !!entry.tags?.includes(tag);
 
 export interface SubAPI {
   /**
@@ -142,76 +117,21 @@ export interface SubAPI {
     elementId?: string,
     options?: boolean | { forceFocus?: boolean; select?: boolean; poll?: boolean }
   ) => boolean | Promise<boolean>;
-
-  /** Resets tag filters in the sidebar to the default filters. */
-  resetTagFilters(): void;
-  /**
-   * Replaces all tag filters in the sidebar with the provided included and excluded lists.
-   *
-   * @param included The tags to include in the filtered stories list
-   * @param excluded The tags to filter out (exclude) from the stories list
-   */
-  setAllTagFilters(included: Tag[], excluded: Tag[]): void;
-  /**
-   * Adds tag filters to the included or excluded filter lists. Included filters are included in the
-   * stories list, whereas excluded filters are filtered out.
-   *
-   * @param tags The tags to add as filters.
-   * @param excluded Whether to add the tags to the include or exclude filter list.
-   */
-  addTagFilters(tags: Tag[], excluded: boolean): void;
-  /**
-   * Removes tag filters from both the included and excluded filter lists.
-   *
-   * @param tags The tags to remove from filters.
-   */
-  removeTagFilters(tags: Tag[]): void;
-  /** Gets the function to use to filter the index based on a given tag. */
-  getFilterFunction(tag: Tag): FilterFunction | null;
-  /** Gets the default included tag filters. */
-  getDefaultIncludedTagFilters(): Tag[];
-  /** Gets the default excluded tag filters. */
-  getDefaultExcludedTagFilters(): Tag[];
-  /** Gets the currently included tag filters. */
-  getIncludedTagFilters(): Tag[];
-  /** Gets the currently excluded tag filters. */
-  getExcludedTagFilters(): Tag[];
 }
 
 type PartialSubState = Partial<SubState>;
-
-const getDefaultTagsFromPreset = memoize(1)((
-  presets: TagsOptions
-): { included: Tag[]; excluded: Tag[] } => {
-  const presetEntries = Object.entries(presets);
-  return {
-    included: presetEntries
-      .filter(([, option]) => option.defaultFilterSelection === 'include')
-      .map(([tag]) => tag),
-    excluded: presetEntries
-      .filter(([, option]) => option.defaultFilterSelection === 'exclude')
-      .map(([tag]) => tag),
-  };
-});
 
 export const DEFAULT_NAV_SIZE = 300;
 export const DEFAULT_BOTTOM_PANEL_HEIGHT = 300;
 export const DEFAULT_RIGHT_PANEL_WIDTH = 400;
 
 export const getDefaultLayoutState: () => SubState = () => {
-  // tagPresets is a local copy of global.TAGS_OPTIONS. Neither is expected to change at runtime.
-  const tagPresets = global.TAGS_OPTIONS || {};
-  const defaultTags = getDefaultTagsFromPreset(tagPresets);
-
   return {
     ui: {
       enableShortcuts: true,
     },
     layout: {
       initialActive: ActiveTabs.CANVAS,
-      tagPresets,
-      includedTagFilters: defaultTags.included,
-      excludedTagFilters: defaultTags.excluded,
       showToolbar: true,
       navSize: DEFAULT_NAV_SIZE,
       bottomPanelHeight: DEFAULT_BOTTOM_PANEL_HEIGHT,
@@ -270,41 +190,7 @@ const getRecentVisibleSizes = (layoutState: API_Layout) => {
   };
 };
 
-const recomputeFilters = (fullAPI: Parameters<ModuleFn>[0]['fullAPI'], store: Store) => {
-  const {
-    layout: { includedTagFilters, excludedTagFilters },
-  } = store.getState();
-
-  const computeFilterFunctions = (set: Tag[]): FilterFunction[][] => {
-    return Object.values(
-      set.reduce(
-        (acc, tag) => {
-          if (Object.hasOwn(BUILT_IN_FILTERS, tag)) {
-            acc['built-in'].push(BUILT_IN_FILTERS[tag as keyof typeof BUILT_IN_FILTERS]);
-          } else {
-            acc.user.push(USER_TAG_FILTER(tag));
-          }
-          return acc;
-        },
-        { 'built-in': [], user: [] } as { 'built-in': FilterFunction[]; user: FilterFunction[] }
-      )
-    ).filter((group) => group.length > 0);
-  };
-
-  fullAPI.experimental_setFilter?.(TAGS_FILTER, (item: API_PreparedIndexEntry) => {
-    const included = computeFilterFunctions(includedTagFilters);
-    const excluded = computeFilterFunctions(excludedTagFilters);
-
-    return (
-      (!included.length ||
-        included.every((group) => group.some((filterFn) => filterFn(item, false)))) &&
-      (!excluded.length ||
-        excluded.every((group) => group.every((filterFn) => filterFn(item, true))))
-    );
-  });
-};
-
-export const init: ModuleFn<SubAPI, SubState> = ({ fullAPI, store, provider, singleStory }) => {
+export const init: ModuleFn<SubAPI, SubState> = ({ store, provider, singleStory }) => {
   const api = {
     toggleFullscreen(nextState?: boolean) {
       return store.setState(
@@ -662,110 +548,6 @@ export const init: ModuleFn<SubAPI, SubState> = ({ fullAPI, store, provider, sin
         store.setState({ theme: updatedTheme });
       }
     },
-
-    getDefaultIncludedTagFilters: () => {
-      const state = store.getState();
-      const { tagPresets } = state.layout;
-      return getDefaultTagsFromPreset(tagPresets).included;
-    },
-
-    getDefaultExcludedTagFilters: () => {
-      const state = store.getState();
-      const { tagPresets } = state.layout;
-      return getDefaultTagsFromPreset(tagPresets).excluded;
-    },
-
-    getIncludedTagFilters: () => {
-      const state = store.getState();
-      return state.layout.includedTagFilters;
-    },
-
-    getExcludedTagFilters: () => {
-      const state = store.getState();
-      return state.layout.excludedTagFilters;
-    },
-
-    resetTagFilters: async () => {
-      const state = store.getState();
-      const { tagPresets } = state.layout;
-      const { included, excluded } = getDefaultTagsFromPreset(tagPresets);
-      await store.setState(
-        (s: State) => ({
-          layout: {
-            ...s.layout,
-            includedTagFilters: included,
-            excludedTagFilters: excluded,
-          },
-        }),
-        { persistence: 'permanent' }
-      );
-      recomputeFilters(fullAPI, store);
-    },
-
-    setAllTagFilters: async (included: Tag[], excluded: Tag[]) => {
-      await store.setState(
-        (s: State) => ({
-          layout: {
-            ...s.layout,
-            includedTagFilters: included,
-            excludedTagFilters: excluded,
-          },
-        }),
-        { persistence: 'permanent' }
-      );
-      recomputeFilters(fullAPI, store);
-    },
-
-    addTagFilters: async (tags: Tag[], excluded: boolean) => {
-      await store.setState(
-        (s: State) => {
-          const newIncluded = new Set(s.layout.includedTagFilters);
-          const newExcluded = new Set(s.layout.excludedTagFilters);
-          for (const tag of tags) {
-            if (excluded) {
-              newIncluded.delete(tag);
-              newExcluded.add(tag);
-            } else {
-              newIncluded.add(tag);
-              newExcluded.delete(tag);
-            }
-          }
-          return {
-            layout: {
-              ...s.layout,
-              includedTagFilters: Array.from(newIncluded),
-              excludedTagFilters: Array.from(newExcluded),
-            },
-          };
-        },
-        { persistence: 'permanent' }
-      );
-      recomputeFilters(fullAPI, store);
-    },
-
-    removeTagFilters: async (tags: Tag[]) => {
-      await store.setState(
-        (s: State) => {
-          return {
-            layout: {
-              ...s.layout,
-              includedTagFilters: s.layout.includedTagFilters.filter((tag) => !tags.includes(tag)),
-              excludedTagFilters: s.layout.excludedTagFilters.filter((tag) => !tags.includes(tag)),
-            },
-          };
-        },
-        { persistence: 'permanent' }
-      );
-      recomputeFilters(fullAPI, store);
-    },
-
-    getFilterFunction(tag: Tag): FilterFunction | null {
-      if (Object.hasOwn(BUILT_IN_FILTERS, tag)) {
-        return BUILT_IN_FILTERS[tag as keyof typeof BUILT_IN_FILTERS];
-      } else {
-        return USER_TAG_FILTER(tag);
-      }
-    },
   };
 
   const persisted = pick(store.getState(), ['layout', 'selectedPanel']);
@@ -774,15 +556,8 @@ export const init: ModuleFn<SubAPI, SubState> = ({ fullAPI, store, provider, sin
     api.setOptions(merge(api.getInitialOptions(), persisted));
   });
 
-  provider.channel?.on(STORY_INDEX_INVALIDATED, () => {
-    recomputeFilters(fullAPI, store);
-  });
-
   return {
     api,
     state: merge(api.getInitialOptions(), persisted),
-    init: () => {
-      recomputeFilters(fullAPI, store);
-    },
   };
 };
