+++ code/core/src/types/modules/api.ts
@@ -12,7 +12,7 @@ import type {
   API_PreparedIndexEntry,
 } from './api-stories';
 import type { SetStoriesStory, SetStoriesStoryData } from './channelApi';
-import type { DocsOptions, TagsOptions } from './core-common';
+import type { DocsOptions } from './core-common';
 import type { StoryIndex } from './indexer';
 
 type OrString<T extends string> = T | (string & {});
@@ -93,12 +93,6 @@ export interface API_Layout {
   panelPosition: API_PanelPositions;
   showTabs: boolean;
   showToolbar: boolean;
-  /** Initial tag filters applied when Storybook loads (not accounting for persisted store state). */
-  tagPresets: TagsOptions;
-  /** Tags to include in the filter (entries with this tag are shown). Persisted permanently. */
-  includedTagFilters: string[];
-  /** Tags to exclude from the filter (entries with this tag are hidden). Persisted permanently. */
-  excludedTagFilters: string[];
 }
 
 export interface API_LayoutCustomisations {
